Rewritten sample function for SQLite FTS. Corrected tests.
