+++ Tests/Linq/Linq/FunctionTests.cs
@@ -5,7 +5,6 @@ using System.Linq.Expressions;
 
 using LinqToDB;
 using LinqToDB.Linq;
-using LinqToDB.SqlQuery;
 using NUnit.Framework;
 
 // ReSharper disable UnusedMember.Local
@@ -453,19 +452,6 @@ namespace Tests.Linq
 					select p);
 		}
 
-		[Test, IncludeDataContextSource(true, ProviderName.SQLite)]
-		public void MatchFtsTest(string context)
-		{
-			using (var db = GetDataContext(context))
-			{
-				var q = from c in db.Types
-					where SqlLite.MatchFts(c, "some*")
-					select c;
-
-				var str = q.ToString();
-				Assert.True(str.Contains(" match "));
-			}
-		}
 	}
 
 	public static class FunctionExtension
@@ -481,34 +467,6 @@ namespace Tests.Linq
 			throw new InvalidOperationException();
 		}
 
-	}
-
-	public static class SqlLite
-	{
-		class MatchBuilder : Sql.IExtensionCallBuilder
-		{
-			public void Build(Sql.ISqExtensionBuilder builder)
-			{
-				var field = builder.GetExpression("src") as SqlField;
-				if (field == null)
-					throw new InvalidOperationException("Can not get table");
-
-				var sqlTable = (SqlTable) field.Table;
-				var newField = new SqlField
-				{
-					Name  = sqlTable.PhysicalName,
-					Table = sqlTable
-				};
-
-				builder.AddParameter("table_field", newField);
-			}
-		}
-
-		[Sql.Extension("{table_field} match {match}", BuilderType = typeof(MatchBuilder), IsPredicate = true)]
-		public static bool MatchFts<TEntity>(TEntity src, [ExprParameter]string match)
-		{
-			throw new InvalidOperationException();
-		}
 	}
 
 }
